Renaming controller to fix naming
